requirements.txt
numpy==1.21.0
onnxruntime==1.9.0
Pillow==9.0.1
streamlit==1.1.0
plotly==5.11.0